Files renamed and automate the return transfer
